Numerical Analysis of numerical methods
GRASP finite difference algorithms for solving the Dirac equations are based on approximating integrations.  Algorithms are to be evaluated by approximating the first derivative and viewing the problem as a linear system of equations.  Also, efficient methods are to be tested for Slater integrals (particularly high Z) and frequency dependent integrals of the transverse-photon correction.